refactor: SelectWeekPage 구성 컴포넌트 이름 변경
